Move to a firefox XPI plugin using JS-Ctypes instead of NPAPI.
JS-Ctypes is intended to make simple calls out to dynamic libraries much simpler to develop than either NPAPI or XPCOMM.  Using this, generate calls to the code contained in ../gssapi_utils and the GSS-API calls themselves to call out to the requisite functionality.

This first commit contains working code that calls out to both libgssapi_krb5 and libgssapi_utils, leading up to a call to create the context token from a call to gss_init_sec_context.  Well, it is "working" in the sense that the calls all succeed in crossing the boundary between JavaScript and dynamic shared object libraries; it fails to invoke the function well enough for the function to succeed.