Message passing with cookies (app_tag, gssweb_bg_tag, etc.)
The messages passed in to the C++ code will now have the following elements copied in to the response:
  cookies
  method

The c++ code was refactored to have a common point for top-level JSON parsing, and the GSSCommand subclasses now only pull out their specific arguments.

GSSCreateSecContextCommand was renamed to GSSInitSecContext.

There was an error in util_json in how a JSONObject implemented operator= that made it at best a no-op.  (At worst it caused crashes.)