Remove a wasted check of XMLHttpResponse state
The state was checked immediately after the send() call - which yields unpredictable results because of the inherent multithreading - and then does nothing more than log a message about it.  So, this winds up logging error messages at times when everything is perfectly fine, but doesn't recover from any sort of error.