Package the Chrome files into a .crx bundle.
Running cmake over the top level directory will now create a set of makefiles that will recurse not only to json_gssapi, but also to browsers/chrome.  As a result, ${build}/browsers/chrome/gssweb.crx is now built.

Note that this requires a file, 'chrome.pem', to exist in the top-level directory.  This file contains encryption keying.